requirements.txt
numpy
torch
pandas
python-dotenv